refactor: remove deprecated pipeline aliases and compat.py module
- Delete application/pipelines/compat.py with deprecated GenericPipeline wrappers
- Remove __getattr__ lazy-loading for PipelineExecutor/RecordProcessor from core/__init__.py
- Update package __init__.py files to import from canonical module locations:
  - chembl/__init__.py: imports from activity.py, assay.py, etc.
  - pubchem/__init__.py: imports from compound.py
  - uniprot/__init__.py: imports from protein.py
  - pubmed/__init__.py: imports from publications.py
- Real pipeline classes remain available at their canonical import paths
- Update CHANGELOG.md with breaking change documentation

BREAKING CHANGE: Deprecated aliases removed. Use direct imports:
- ChEMBLActivityPipeline: from bioetl.application.pipelines.chembl.activity
- PubChemCompoundPipeline: from bioetl.application.pipelines.pubchem.compound
- UniProtProteinPipeline: from bioetl.application.pipelines.uniprot.protein
- PubMedPublicationsPipeline: from bioetl.application.pipelines.pubmed.publications
- PipelineExecutor/RecordProcessor: from bioetl.application.core.executor/record_processor
