+++ src/bioetl/application/pipelines/pubchem/__init__.py
@@ -4,13 +4,13 @@ This package provides pipelines and transformers for extracting and
 processing data from the PubChem database.
 
 Main Components:
+- PubChemCompoundPipeline: Pipeline for compound data
 - PubChemCompoundTransformer: Transformer for compound data
-- PubChemCompoundPipeline: Deprecated alias (use GenericPipeline)
 """
 
 from __future__ import annotations
 
-from bioetl.application.pipelines.compat import PubChemCompoundPipeline
+from bioetl.application.pipelines.pubchem.compound import PubChemCompoundPipeline
 from bioetl.application.pipelines.pubchem.transformer import PubChemCompoundTransformer
 
 __all__ = [
