+++ src/bioetl/application/pipelines/uniprot/__init__.py
@@ -4,13 +4,13 @@ This package provides pipelines and transformers for extracting and
 processing data from the UniProt database.
 
 Main Components:
+- UniProtProteinPipeline: Pipeline for protein data
 - UniProtProteinTransformer: Transformer for protein data
-- UniProtProteinPipeline: Deprecated alias (use GenericPipeline)
 """
 
 from __future__ import annotations
 
-from bioetl.application.pipelines.compat import UniProtProteinPipeline
+from bioetl.application.pipelines.uniprot.protein import UniProtProteinPipeline
 from bioetl.application.pipelines.uniprot.transformer import UniProtProteinTransformer
 
 __all__ = [
